+++ codex/serializers/admin/stats.py
@@ -19,8 +19,8 @@ FILE_TYPES_CHOICES = ("CBZ", "CBR", "CBT", "PDF", "UNKNOWN")
 class StatsSystemSerializer(Serializer):
     """Platform System Information."""
 
-    name = CharField(required=False, read_only=True)
-    release = CharField(required=False, read_only=True)
+    name = CharField(read_only=True, required=False)
+    release = CharField(read_only=True, required=False)
 
 
 class StatsPlatformSerializer(Serializer):
@@ -42,7 +42,7 @@ class StatsConfigSerializer(Serializer):
     user_registered_count = IntegerField(required=False, read_only=True)
     auth_group_count = IntegerField(required=False, read_only=True)
     # Only for api
-    api_key = CharField(required=False, read_only=True)
+    api_key = CharField(read_only=True, required=False)
 
 
 class StatsSessionsSerializer(Serializer):
@@ -123,4 +123,4 @@ class AdminStatsRequestSerializer(Serializer):
 class APIKeySerializer(Serializer):
     """API Key."""
 
-    api_key = CharField(source="name")
+    api_key = CharField(source="name", read_only=True)