+++ codex/serializers/browser/settings.py
@@ -9,8 +9,9 @@ from rest_framework.serializers import (
 )
 
 from codex.choices import BROWSER_ORDER_BY_CHOICES
-from codex.serializers.browser.filters import BrowserSettingsFilterSerializer
-from codex.serializers.fields import BreadcrumbsField, TimestampField, TopGroupField
+from codex.serializers.browser.filters import BrowserSettingsFilterInputSerializer
+from codex.serializers.fields import TimestampField
+from codex.serializers.fields.browser import BreadcrumbsField, TopGroupField
 from codex.serializers.route import SimpleRouteSerializer
 
 
@@ -26,7 +27,8 @@ class BrowserSettingsShowGroupFlagsSerializer(Serializer):
 class BrowserFilterChoicesInputSerilalizer(Serializer):
     """Browser Settings for the filter choices response."""
 
-    filters = BrowserSettingsFilterSerializer(required=False)
+    filters = BrowserSettingsFilterInputSerializer(required=False)
+    # NOT Sanitized because so complex.
     q = CharField(allow_blank=True, required=False)