+++ codex/serializers/opds/urls.py
@@ -1,6 +1,7 @@
 """OPDS URLs."""
 
-from rest_framework.serializers import CharField, Serializer
+from rest_framework.fields import CharField
+from rest_framework.serializers import Serializer
 
 
 class OPDSURLsSerializer(Serializer):