+++ codex/serializers/opds/v2/metadata.py
@@ -1,9 +1,6 @@
 """OPDS 2.0 Metadata Serializer."""
 
-from rest_framework.fields import (
-    CharField,
-    DateTimeField,
-)
+from rest_framework.fields import CharField, DateTimeField
 from rest_framework.serializers import Serializer
 
 
@@ -15,9 +12,7 @@ class OPDS2MetadataSerializer(Serializer):
     """
 
     identifier = CharField(read_only=True, required=False)
-    title = CharField(
-        read_only=True,
-    )
+    title = CharField(read_only=True)
     subtitle = CharField(read_only=True, required=False)
     modified = DateTimeField(read_only=True, required=False)
     description = CharField(read_only=True, required=False)